requirements.txt
asgiref==3.8.1
attrs==25.3.0
certifi==2025.6.15
charset-normalizer==3.4.2
Django==5.2.3
django-debug-toolbar==5.2.0
django-filter==25.1
djangorestframework==3.16.0
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.6.1
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
Markdown==3.8.2
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
python-dotenv==1.1.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.25.1
sqlparse==0.5.3
uritemplate==4.2.0
urllib3==2.5.0
whitenoise==6.9.0
